backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
psycopg2-binary
pydantic-settings
greenlet
rapidfuzz
holidays
numpy
httpx
python-dotenv
beautifulsoup4
flask-cors
httpcore